Moved famous quote into its own component to handle animations
