+++ src/DigitalGarden/DigitalGarden.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
-VisualStudioVersion = 18.0.11217.181 d18.0
+VisualStudioVersion = 18.0.11217.181
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalGarden", "DigitalGarden\DigitalGarden\DigitalGarden.csproj", "{43FB2861-211B-465C-B6E8-2A7FBEF1BB45}"
 EndProject
