+++ src/DigitalGarden/DigitalGarden/DigitalGarden.Client/Components/Pages/Home.razor
@@ -1,6 +1,5 @@
 ﻿@page "/"
 @rendermode InteractiveAuto
-@implements IDisposable
 
 <PageTitle>Home</PageTitle>
 
@@ -16,23 +15,13 @@
     
     <div class="grid-12 mt-4 mb-6">
         <div class="quote-card col-12 col-start-m-3 col-m-8 py-3 px-4">
-            @if (QuoteOfTheDay is null || QuoteWords is null)
+            @if (QuoteOfTheDay is null)
             {
                 <PlanetSpinner />
             }
             else
             {
-                <blockquote>
-                    <q>
-                        @for (var i = 0; i < QuoteWords.Length; i++)
-                        {
-                            var word = QuoteWords[i];
-                            var spanClass = ShowWordIndices.Contains(i) ? ShowClassName : string.Empty;
-                            <span class="@spanClass">@word</span>
-                        }
-                    </q>
-                    <cite class="@(ShowWordsCompleted ? ShowClassName : string.Empty)">@QuoteOfTheDay.Author</cite>
-                </blockquote>
+                <AnimatedFamousQuote Quote="QuoteOfTheDay" />
             }            
         </div>
     </div>
